nexiu-ops-poc/requirements.txt
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.25.0
pandas>=2.1.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
